Hackathon_0/platinum/requirements.txt
# Platinum Tier: AI Employee Requirements
# Python 3.10+ required
# Install: pip install -r requirements.txt

# ===========================================
# CORE DEPENDENCIES
# ===========================================
python-dotenv==1.0.0
colorlog==6.8.0
PyYAML==6.0.1

# ===========================================
# GMAIL API (REQUIRED for email sending)
# ===========================================
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

# ===========================================
# GIT SYNC (Vault synchronization)
# ===========================================
GitPython==3.1.41

# ===========================================
# HTTP & API CLIENTS
# ===========================================
requests==2.31.0
aiohttp==3.9.1

# ===========================================
# SOCIAL MEDIA APIs
# ===========================================
# Facebook Graph API
facebook-sdk==3.1.0

# Twitter API v2
tweepy==4.14.0

# ===========================================
# AI CONTENT GENERATION (Gemini - FREE!)
# ===========================================
google-generativeai>=0.7.0

# ===========================================
# ODOO INTEGRATION (Accounting)
# ===========================================
# XML-RPC client for Odoo JSON-RPC
# (Built into Python standard library)

# ===========================================
# WHATSAPP (Local only - from Gold tier)
# ===========================================
playwright==1.40.0

# ===========================================
# WEB AUTOMATION
# ===========================================
selenium==4.16.0
beautifulsoup4==4.12.2

# ===========================================
# DATA PROCESSING
# ===========================================
pandas==2.1.4
python-dateutil==2.8.2

# ===========================================
# FILE MONITORING
# ===========================================
watchdog==3.0.0

# ===========================================
# UTILITIES
# ===========================================
markdown==3.5.1
jinja2==3.1.2
pydantic==2.5.3

# ===========================================
# HEALTH MONITORING
# ===========================================
flask==3.0.0
psutil==5.9.7
docker==7.0.0

# ===========================================
# LOGGING
# ===========================================
python-json-logger==2.0.7

# ===========================================
# SECURITY
# ===========================================
cryptography==41.0.7

# ===========================================
# CLOUD DEPLOYMENT (Optional)
# ===========================================
# For Oracle Cloud/AWS deployment scripts
paramiko==3.4.0
scp==0.14.5

# ===========================================
# TESTING (Optional - for development)
# ===========================================
# pytest==7.4.4
# pytest-asyncio==0.23.3
